Add module to represent a bibliography
The new `bibliography` module provides the `Bibliography` class which
represents a collection of `BibliographyEntry`s. The `Bibliography`
class takes an adapter which is the gateway to the backend that knows
how to construct this collection from some data source.

The `BibliographyAdapter` abstract class is implemented for the case
that the bibliography is defined by a Bibtex file on disk. The
implementation uses the `bibtexparser` library to parse the file and
converts the parsed entries into `BibliographicEntry` instances.