Integrate the `Bibliography` with the index template
The `BiblaryIndexView` is updated to add the `entries` field to the
context. This is a list of `BibliographyEntry` objects that the template
can then display. The entries are obtained from a `Bibliography` which
is constructed in the `get_bibliography` method, passing the required
adapter as a dependency.

The adapter implementation and its configuration are obtained from the
package settings as they need to be configurable. The package settings
are defined and managed through the `biblary.settings.Settings` class.
It will automatically prepend a prefix to the settings to namespace them
to this application. In this case that is `BIBLARY` and so, for example,
to set the `bibliography_adapter` setting in the global configuration
one would define the constant `BIBLARY_BIBLIOGRAPHY_ADAPTER`.