Add an interface and implementation for a file storage
To provide the option to serve files alongside bibliographic entries, a
new optional dependency is injected into the `Bibliography` class under
the argument `storage`. It takes an instance of `AbstractStorage` which
is a interface to a simple file store.

The `AbstractStorage` is implemented as `FileSystemStorage` which will
simply store the files on the local file system. Each bibliographic
entry gets its own sub directory, the name of which is the sha256 hash
of the identifier. The files are stored within this directory where the
filename is the value of the `FileType` enumeration. This design
directly results that each bibliographic entry can only store a maximum
of one files for each of the types and only for those types.

This particular design may very well turn out too limiting in the future
but at least for now this approach allows to automatically discover what
files exist for an entry without requiring a dedicated index that keeps
track of this. Note that for the local file system solution, one could
of course discover what files exists by simply listing the contents of
the entry's folder, however, other stores, such as an object store, may
not support this kind of discovery and require an explicit key to be
known. By limiting what files can be written through the `FileType` enum
one can now just iterate over its values and call `exists` to determine
whether that file type exists for an entry.