Integrate the file storage functionality into the index template
Two new settings are added that allow to configure the file storage:

 * `BIBLIOGRAPHY_STORAGE`
 * `BIBLIOGRAPHY_STORAGE_CONFIGURATION`

These optional settings, when specified, will cause the `Bibliography`
when it is initialized to be configured with the storage constructed.
If the bibliography has a configured storage, the index view will add a
mapping to each bibliographic entry in the context that maps the known
file types (defined by the `FileType` enum) to a boolean, indicating
whether the entry has a file stored of the corresponding type.

This additional information allows the template to construct links to
download the content of the files contained within the configured
storage. The content is served by the new `BiblaryFileView` which is
routed behind the `/file` URL. It takes the entry identifier and file
type as URL parameters and when they are valid, the byte content will be
served as a PDF. This file type is currently hardcoded but may need to
become dynamic in the future if the file storage will allow to store
other file types.