`AbstractStorage`: add the `put_file` method to write a file
The new `put_file` allows to write binary content to a file for a
specific bibliographic entry and file type. It is implemented for the
`FileSystemStorage` implementation. Currently there are no checks
whether the file already exists but it is simply overwritten.

This new method allows to remove the ad-hoc method that was added to the
`Bibliography` instance returned by the `get_bibliography` fixture to
allow to write files to the bibliography. The one test that was using it
now simply uses the new `put_file` interface.