Add view to upload a file to the configure storage
The last commit introduced the `put_file` method to the `AbstractStorage`
interface which allows to write a file of a certain type for an entry of
the bibliography. Here this functionality is plumbed in through the
`BibliographyUploadView` which is routed by default to the `upload` URL.

The view uses the `BibliographyUploadFileForm` which contains 3 fields:

 * `entry_identifier
 * `file_type`
 * `content`

where the first are choice fields and the last is a file upload field.
The choices of the two choice fields are dynamically loaded from the
contents of the configured bibliography and the values of the `FileType`
enumeration, respectively.