Add template tag to mark author span in template as main author
The new template tag `biblary.templatetags.authors.main_author_class`
will take the name of an author and compare it to a sequence of regex
patterns. If any of it matches, it will return the string that is
configured for the `BIBLIOGRAPHY_MAIN_AUTHOR_CLASS` setting. The
patterns can be configured by the `BIBLIOGRAPHY_MAIN_AUTHOR_PATTERNS`
setting.

This mechanism can be used in the index template to mark certain authors
with an additional CSS class, which in turn allows them to be styled
differently, such that they stand out more.