`Bibliography`: add the `add_entry` method.
This method allows to add a new entry to the bibliography. It accepts a
`BibliographicEntry` or a string, in which case the `parse_entry` method
of the adapter is called first to attempt to parse it into a
`BibliographicEntry`.

The method returns the entry that was added, which is useful in case the
argument was of string type and the entry was parsed by the adapter.
Note that the new state of the bibliography is not persisted to the
original source of the adapter.