Add view to download entry formatted as bibtex
The content of the bibliographic entry is formatted in bibtex and
provided as a plain text attachment in the HTTP response.

The view directly uses the `BibtexBibliography` class which has a new
class method `write_entry` that writes the content of the entry
formatted as bibtex to a provided text stream.